grpc-client/requirements.txt
blinker==1.9.0
click==8.2.1
colorama==0.4.6
Flask==3.1.2
grpcio==1.74.0
grpcio-tools==1.74.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
protobuf==6.32.0
setuptools==80.9.0
Werkzeug==3.1.3
